define signaling handler on server
